Switch to Hyper 0.12 & Add a previewabel example
I switch back to Hyper is because at least I can find some examples
or documentation about how to create a service using it, but actix
doesn't. Actix is also too heavy for this use case.

The basic structure has been confirmed. Now we have a "working"
example of this structure. I can create and register a hook, create
a service, and use `hyper::rt::run` to serve it. As you can see in
the cousole output, registered hook will be executed when needed.
This at least prooves that the mechanism of hook-registering,
finding the right hook and Hyper takes my struct.

I don't know why I say this much in the commit message, emmmmmmmmm.

Nobody cares.

Signed-off-by: Kay Lin <kaymw@aosc.io>