Remove dummy code & it actually works
It works under both `application/json` and
`application/x-www-form-urlencoded` mode, which is the main cause
for me to add another filed to the `Delivery` struct.

Right now it's ugly but it can work, I've tested it on Github.

So I need to clean it up. add some comments and then it will ready
to be published.

Signed-off-by: Kay Lin <kaymw@aosc.io>